Move HabitWithEntries to domain.model package
